move function definition into a single file
