--- functions/create_issue/definition.ts
@@ -1,48 +1,0 @@
-import { DefineFunction, Schema } from "deno-slack-sdk/mod.ts";
-
-const CreateIssueDefinition = DefineFunction({
-  callback_id: "create_issue",
-  title: "Create GitHub issue",
-  description: "Create a new GitHub issue in a repository",
-  source_file: "functions/create_issue/mod.ts",
-  input_parameters: {
-    properties: {
-      githubAccessTokenId: {
-        type: Schema.slack.types.oauth2,
-        oauth2_provider_key: "github",
-      },
-      url: {
-        type: Schema.types.string,
-        description: "Repository URL",
-      },
-      title: {
-        type: Schema.types.string,
-        description: "Issue Title",
-      },
-      description: {
-        type: Schema.types.string,
-        description: "Issue Description",
-      },
-      assignees: {
-        type: Schema.types.string,
-        description: "Assignees",
-      },
-    },
-    required: ["githubAccessTokenId", "url", "title"],
-  },
-  output_parameters: {
-    properties: {
-      GitHubIssueNumber: {
-        type: Schema.types.number,
-        description: "Issue number",
-      },
-      GitHubIssueLink: {
-        type: Schema.types.string,
-        description: "Issue link",
-      },
-    },
-    required: ["GitHubIssueNumber", "GitHubIssueLink"],
-  },
-});
-
-export default CreateIssueDefinition;
