+++ functions/create_issue_test.ts
@@ -1,7 +1,7 @@
 import * as mf from "https://deno.land/x/mock_fetch@0.3.0/mod.ts";
 import { SlackFunctionTester } from "deno-slack-sdk/mod.ts";
 import { assertEquals } from "https://deno.land/std@0.153.0/testing/asserts.ts";
-import handler from "./mod.ts";
+import handler from "./create_issue.ts";
 
 // Replaces globalThis.fetch with the mocked copy
 mf.install();
@@ -24,9 +24,11 @@ const env = { logLevel: "CRITICAL" };
 
 Deno.test("Create a GitHub issue with given inputs", async () => {
   const inputs = {
-    url: "https://github.com/slack-samples/deno-github-functions",
-    title: "The issue title",
     githubAccessTokenId: {},
+    url: "https://github.com/slack-samples/deno-github-functions",
+    githubIssue: {
+      title: "The issue title",
+    },
   };
   const { outputs } = await handler(createContext({ inputs, env }));
   assertEquals(outputs?.GitHubIssueNumber, 123);
